% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.genealogy-package.R
\docType{package}
\name{maths.genealogy-package}
\alias{maths.genealogy}
\alias{maths.genealogy-package}
\title{maths.genealogy: Mathematics Genealogy Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Query, extract, and plot genealogical data from The Mathematics Genealogy Project \url{https://mathgenealogy.org/}. Data is gathered from the WebSocket server run by the 'geneagrapher-core' project \url{https://github.com/davidalber/geneagrapher-core}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/louisaslett/maths.genealogy}
  \item Report bugs at \url{https://github.com/louisaslett/maths.genealogy/issues}
}

}
\author{
\strong{Maintainer}: Louis Aslett \email{louis.aslett@durham.ac.uk} (\href{https://orcid.org/0000-0003-2211-233X}{ORCID})

}
\keyword{internal}
