% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grviz.R
\name{plot_grviz}
\alias{plot_grviz}
\title{Plot genealogical tree with Graphviz}
\usage{
plot_grviz(g, file = "")
}
\arguments{
\item{g}{an object of class \code{genealogy}, as returned by \code{\link[=get_genealogy]{get_genealogy()}}.}

\item{file}{an optional file name.
If the file name is specified, then Graphviz will render the genealogical tree to PDF and save in this file.
If the file name is not specified, then the plot will be rendered interactively in the RStudio Viewer panel.}
}
\value{
If a filename was specified, the full path of the saved file is returned as a \code{character(1)} string.
If no filename was specified, then an \code{htmlwidget} suitable for display in the RStudio Viewer is returned.
}
\description{
Plots a genealogical tree either interactively or to PDF using the Graphviz layout engine.
}
\details{
This function requires the \code{DiagrammeR}, \code{DiagrammeRsvg} and \code{svgPanZoom} packages to be installed.
They are only "Suggests" dependencies this package supports multiple plotting.
The presence of these packages will be verified when the function is actually called, providing an opportunity to install them automatically if needed.
}
\examples{
# TODO
}
