% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_id.R
\name{search_id}
\alias{search_id}
\title{Search for mathematician in Mathematics Genealogy Project}
\usage{
search_id(
  family = NULL,
  given = NULL,
  middle = NULL,
  university = NULL,
  year = NULL,
  thesis_keyword = NULL,
  country = NULL,
  discipline = NULL
)
}
\arguments{
\item{family}{a character(1) string with the family names.}

\item{given}{a character(1) string with the given names.}

\item{middle}{a character(1) string with the collapsed middle name(s).}

\item{university}{a character(1) string with the University at which PhD studied.}

\item{year}{a character(1) string or integer(1) with the year of completion.}

\item{thesis_keyword}{a character(1) string with keyword(s) in the PhD thesis title.}

\item{country}{a character(1) string with the country of study.}

\item{discipline}{an integer(1) with the mathematical sub-discipline code.}
}
\value{
Data frame containing all matches against the provided search terms, with columns:
\describe{
\item{\code{id}}{Mathematician ID (as required by \code{\link[=get_genealogy]{get_genealogy()}});}
\item{\code{name}}{The full name (surname first) of the mathematician;}
\item{\code{university}}{The institution at which PhD was obtained;}
\item{\code{year}}{The year PhD was completed.}
}
}
\description{
Perform an online search using information about an individual mathematician to find their ID in the \href{https://mathgenealogy.org/}{Mathematics Genealogy Project}.
}
\details{
Any one or more of the listed arguments can be provided.
This will trigger an online search against the live \href{https://mathgenealogy.org/}{Mathematics Genealogy Project} database, so please be considerate and do not spam queries.
All the information returned by a standard search on the website is gathered into a data frame and returned, enabling programmatic access to the data.

If you cannot find the individual you are looking for, it could be that they are not in the \href{https://mathgenealogy.org/}{Mathematics Genealogy Project} database.
New data can be submitted by following the instructions in the "How to submit updates" section at \url{https://mathgenealogy.org/submit.php}.
}
\examples{
# Search for the package author
ids <- search_id("Aslett", "Louis")

# Then use this to fetch genealogy (just descendents for speed)
# TODO
}
