% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genealogy.R
\name{get_genealogy}
\alias{get_genealogy}
\title{Retrieve genealogy tree by mathematician ID}
\usage{
get_genealogy(id, ancestors = TRUE, descendents = TRUE)
}
\arguments{
\item{id}{vector of integer IDs of mathematicians for whom the genealogy should be retrieved}

\item{ancestors}{logical indicating whether to include the genealogy backward to include all ancestors, defaults to \code{TRUE}.}

\item{descendents}{logical indicating whether to include the genealogy forward to include all descendents, defaults to \code{TRUE}.}
}
\value{
A joint genealogy object
}
\description{
Queries the genealogy of a set of mathematicians by their ID in the \href{https://mathgenealogy.org/}{Mathematics Genealogy Project}
}
\examples{
# TODO
}
